bp: cbr_handleCrs(): Remove unnecessary cast
I'm guessing that commit 5e0dd6d72 (bpv7: persist CRS reception history;
add l crslog / m crsmaxlog commands, 2026-06-02) got through our CI
because our runners only use LP64 and GCC. In that patch, the size_t
returned from sdr_list_length() is compared to the unsigned long
CbrDb.crsHistoryMax which was explicitly casted to long.

CbrDb.crsHistoryMax is unsigned int. When INT_WIDTH < LONG_WIDTH,
casting unsigned int to long, then casting it to unsigned long (because
we're comparing with size_t) is safe. The value remains the same, so GCC
doesn't emit the warning. In contrast, Clang doesn't seem to do that
much analysis and will emit the -Wsign-compare warning which makes the
build fail because of -Werror.

Note that when INT_WIDTH == LONG_WIDTH, casting unsigned int to long is
problematic. If the unsigned value does not fit into a long, the binary
representation is reinterpreted as a negative value. When we compare it
to size_t, signed long undergoes sign-extension to match the width of
unsigned long long, before being converted to an unsigned type, and then
the new value will be different from the original unsigned int value.
Had we compiled the same code, say in an LLP64 environment, then GCC
would have errored on -Wsign-compare as well.

Commit 8e7afd527 (bpv7/cbr: fix macOS clang sign-compare build error in
cbr_handleCrs, 2026-06-10) addressed Clang's warning by casting to
size_t instead of long, but we don't even need a cast. Comparing size_t
and unsigned int like this should be harmless since the narrower
unsigned type would just be promoted to the wider unsigned type.

+++ bpv7/library/cbr.c
@@ -3208,7 +3208,7 @@ int	cbr_handleCrs(Sdr sdr, unsigned char *adminRecord, int length,
 				}
 				else if (cbrDb.crsHistoryMax > 0
 					&& sdr_list_length(sdr, cbrDb.crsHistory)
-						> (size_t) cbrDb.crsHistoryMax)
+						> cbrDb.crsHistoryMax)
 				{
 					/*	Evict oldest entry.	*/
 					firstElt = sdr_list_first(sdr,
